move bert classes depedent on texar inside the  folder
